+++ packages/shared/lib/keys.dart
@@ -1,0 +1,4 @@
+/// Export of the query parameter keys.
+library keys;
+
+export 'src/keys.dart' show QueryParameterKeys;